Secure task CRUD with JWT ownership checks

+++ TaskManagerApi/TaskManagerApi.Application/UseCases/TaskItems/DeleteTask/DeleteTaskUseCase.cs
@@ -1,5 +1,5 @@
-﻿using TaskManagerApi.Application.Commands;
-using TaskManagerApi.Application.Commands.TaskItem;
+﻿using TaskManagerApi.Application.Commands.TaskItem;
+using TaskManagerApi.Application.Exceptions;
 using TaskManagerApi.Application.Interfaces;
 
 namespace TaskManagerApi.Application.UseCases.TaskItems.DeleteTask;
@@ -18,7 +18,10 @@ public class DeleteTaskUseCase
         var task = await _repository.GetTaskByIdAsync(command.Id);
 
         if (task == null)
-            throw new KeyNotFoundException("Task not found");
+            throw new KeyNotFoundException($"Task {command.Id} not found");
+
+        if (task.UserId != command.CurrentUserId)
+            throw new ForbiddenException("You do not have access to this task");
         
         await _repository.DeleteTaskAsync(command.Id);
     }
